+++ src/lr_fhss_v1_base_types.h
@@ -1,7 +1,7 @@
-/*!
- * @file      lr1110_driver_version.h
+/**
+ * @file      lr_fhss_v1_base_types.h
  *
- * @brief     Placeholder to keep the version of LR1110 driver.
+ * @brief     Radio-independent LR-FHSS base type definitions, version 1
  *
  * The Clear BSD License
  * Copyright Semtech Corporation 2021. All rights reserved.
@@ -32,17 +32,16 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef LR1110_DRIVER_VERSION_H
-#define LR1110_DRIVER_VERSION_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
+#ifndef LR_FHSS_V1_BASE_TYPES_H__
+#define LR_FHSS_V1_BASE_TYPES_H__
 
 /*
  * -----------------------------------------------------------------------------
  * --- DEPENDENCIES ------------------------------------------------------------
  */
+
+#include <stdint.h>
+#include <stdbool.h>
 
 /*
  * -----------------------------------------------------------------------------
@@ -54,37 +53,75 @@ extern "C" {
  * --- PUBLIC CONSTANTS --------------------------------------------------------
  */
 
-#define LR1110_DRIVER_VERSION_MAJOR 7
-#define LR1110_DRIVER_VERSION_MINOR 0
-#define LR1110_DRIVER_VERSION_PATCH 0
-
 /*
  * -----------------------------------------------------------------------------
  * --- PUBLIC TYPES ------------------------------------------------------------
  */
+
+/**
+ * @brief LR-FHSS modulation type
+ */
+typedef enum lr_fhss_v1_modulation_type_e
+{
+    LR_FHSS_V1_MODULATION_TYPE_GMSK_488 = 0,
+} lr_fhss_v1_modulation_type_t;
+
+/**
+ * @brief LR-FHSS coding rate
+ */
+typedef enum lr_fhss_v1_cr_e
+{
+    LR_FHSS_V1_CR_5_6 = 0x00,
+    LR_FHSS_V1_CR_2_3 = 0x01,
+    LR_FHSS_V1_CR_1_2 = 0x02,
+    LR_FHSS_V1_CR_1_3 = 0x03,
+} lr_fhss_v1_cr_t;
+
+/**
+ * @brief LR-FHSS grid
+ */
+typedef enum lr_fhss_v1_grid_e
+{
+    LR_FHSS_V1_GRID_25391_HZ = 0x00,
+    LR_FHSS_V1_GRID_3906_HZ  = 0x01,
+} lr_fhss_v1_grid_t;
+
+/**
+ * @brief LR-FHSS bandwidth
+ */
+typedef enum lr_fhss_v1_bw_e
+{
+    LR_FHSS_V1_BW_39063_HZ   = 0x00,
+    LR_FHSS_V1_BW_85938_HZ   = 0x01,
+    LR_FHSS_V1_BW_136719_HZ  = 0x02,
+    LR_FHSS_V1_BW_183594_HZ  = 0x03,
+    LR_FHSS_V1_BW_335938_HZ  = 0x04,
+    LR_FHSS_V1_BW_386719_HZ  = 0x05,
+    LR_FHSS_V1_BW_722656_HZ  = 0x06,
+    LR_FHSS_V1_BW_773438_HZ  = 0x07,
+    LR_FHSS_V1_BW_1523438_HZ = 0x08,
+    LR_FHSS_V1_BW_1574219_HZ = 0x09,
+} lr_fhss_v1_bw_t;
+
+/**
+ * @brief LR-FHSS parameter structure
+ */
+typedef struct lr_fhss_v1_params_s
+{
+    const uint8_t*               sync_word; /**< 4-byte sync word */
+    lr_fhss_v1_modulation_type_t modulation_type;
+    lr_fhss_v1_cr_t              cr;
+    lr_fhss_v1_grid_t            grid;
+    lr_fhss_v1_bw_t              bw;
+    bool                         enable_hopping;
+    uint8_t                      header_count; /**< Number of header blocks */
+} lr_fhss_v1_params_t;
 
 /*
  * -----------------------------------------------------------------------------
  * --- PUBLIC FUNCTIONS PROTOTYPES ---------------------------------------------
  */
 
-/*!
- * @brief Compare version information with current ones
- *
- * This macro expands to true boolean value if the version information provided in argument is compatible or
- * retro-compatible with the version of this code base
- */
-#define LR1110_DRIVER_VERSION_CHECK( x, y, z ) \
-    ( x == LR1110_DRIVER_VERSION_MAJOR &&      \
-      ( y < LR1110_DRIVER_VERSION_MINOR ||     \
-        ( y == LR1110_DRIVER_VERSION_MINOR && z <= LR1110_DRIVER_VERSION_PATCH ) ) )
-
-const char* lr1110_driver_version_get_version_string( void );
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif  // LR1110_DRIVER_VERSION_H
+#endif  // LR_FHSS_V1_BASE_TYPES_H__
 
 /* --- EOF ------------------------------------------------------------------ */