+++ CHANGELOG.md
@@ -4,18 +4,27 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v7.0.0] 2021-12-06
+
+### Added
+
+* [radio] Support of LR-FHSS (see lr1110_lr_fhss.c and lr1110_lr_fhss.h)
+* [system] `lr1110_system_calibrate_image_in_mhz()` function
+
+### Removed
+
+* [GNSS] `lr1110_gnss_set_xtal_error()` and `lr1110_gnss_set_xtal_error()` functions
+
 ## [v6.0.0] 2021-09-24
 
 ### Added
 
 * [GNSS] `LR1110_GNSS_SCAN_MODE_3_SINGLE_SCAN_AND_5_FAST_SCANS` entry `lr1110_gnss_scan_mode_t`
 
-
 ### Changed
 
 * [GNSS] `lr1110_gnss_set_scan_mode()` function does not take out parameter anymore and is now based on lr1110_hal_write()
 * [GNSS] `lr1110_gnss_get_detected_satellites()` function returns also the doppler per satellite
-
 
 ### Removed